Split presentation tests into Client and Server
